+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+* BATS tests
+
 ## [1.1.0] - 2022-04-14
 
 ### Added
-* Merge strategy options for the `ort` strategy. 
+* Merge strategy options for the `ort` strategy.
 
 ## [1.0.0] - 2022-04-14